update dashboard test
