+++ dashboard/Makefile
@@ -4,11 +4,11 @@ venv:
 build:
 	. venv/bin/activate; pip-compile -o requirements.txt; pip install -r requirements.txt
 
-test:
+test: build
 	. venv/bin/activate; export PYTHONPATH=./src:${PYTHONPATH}; pytest -o log_cli=true tests
 
 check:
 	black src tests
 
-run-driver:
+run-driver: build
 	. venv/bin/activate; python src/driver.py
